Add implementation of POP instruction

+++ src/execute/esl_context.cpp
@@ -3,7 +3,7 @@
 esl_context::esl_context()
 {
     this->variables = new std::unordered_map<std::string, esl_variable *>;
-    this->functions = new std::unordered_map<std::string, int>;
+    this->functions = new std::unordered_map<std::string, size_t>;
     this->pc = 0;
 }
 
@@ -18,12 +18,12 @@ esl_variable *esl_context::get_variable(std::string var_name)
     return this->variables->at(var_name);
 }
 
-int esl_context::get_function(std::string fun_name)
+size_t esl_context::get_function(std::string fun_name)
 {
     return this->functions->at(fun_name);
 }
 
-int esl_context::get_pc()
+size_t esl_context::get_pc()
 {
     return this->pc;
 }
@@ -33,12 +33,12 @@ void esl_context::set_variable(const std::string &var_name, esl_variable *value)
     (*this->variables)[var_name] = value;
 }
 
-void esl_context::set_function(const std::string &fun_name, int value)
+void esl_context::set_function(const std::string &fun_name, size_t value)
 {
     (*this->functions)[fun_name] = value;
 }
 
-void esl_context::set_pc(int pc)
+void esl_context::set_pc(size_t pc)
 {
     this->pc = pc;
 }
